Move index to commons
