+++ src/main/java/spleetwaise/address/logic/commands/RemarkCommand.java
@@ -6,10 +6,10 @@ import static spleetwaise.commons.util.CollectionUtil.requireAllNonNull;
 
 import java.util.List;
 
-import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.Messages;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.model.person.Remark;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.commands.Command;
 import spleetwaise.commons.logic.commands.CommandResult;
 import spleetwaise.commons.logic.commands.exceptions.CommandException;
