+++ src/main/java/spleetwaise/address/logic/parser/RemarkCommandParser.java
@@ -4,9 +4,9 @@ import static java.util.Objects.requireNonNull;
 import static spleetwaise.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static spleetwaise.address.logic.parser.CliSyntax.PREFIX_REMARK;
 
-import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.commands.RemarkCommand;
 import spleetwaise.address.model.person.Remark;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.exceptions.IllegalValueException;
 import spleetwaise.commons.logic.parser.Parser;
 import spleetwaise.commons.logic.parser.exceptions.ParseException;
