+++ src/main/java/spleetwaise/transaction/logic/parser/AddCommandParser.java
@@ -11,11 +11,11 @@ import static spleetwaise.transaction.model.transaction.Date.getNowDate;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.parser.ArgumentMultimap;
 import spleetwaise.address.logic.parser.ArgumentTokenizer;
 import spleetwaise.address.logic.parser.Prefix;
 import spleetwaise.address.model.person.Person;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.parser.Parser;
 import spleetwaise.commons.logic.parser.exceptions.ParseException;
 import spleetwaise.transaction.logic.commands.AddCommand;
