+++ src/main/java/spleetwaise/transaction/logic/parser/EditCommandParser.java
@@ -15,11 +15,11 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.parser.ArgumentMultimap;
 import spleetwaise.address.logic.parser.ArgumentTokenizer;
 import spleetwaise.address.logic.parser.Prefix;
 import spleetwaise.address.model.person.Phone;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.parser.Parser;
 import spleetwaise.commons.logic.parser.exceptions.ParseException;
 import spleetwaise.transaction.logic.commands.EditCommand;
