+++ src/main/java/spleetwaise/transaction/logic/parser/UnmarkCommandParser.java
@@ -2,7 +2,7 @@ package spleetwaise.transaction.logic.parser;
 
 import static spleetwaise.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
-import spleetwaise.address.commons.core.index.Index;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.parser.Parser;
 import spleetwaise.commons.logic.parser.exceptions.ParseException;
 import spleetwaise.transaction.logic.commands.UnmarkCommand;
