+++ src/test/java/spleetwaise/address/logic/commands/CommandTestUtil.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.parser.CliSyntax;
 import spleetwaise.address.model.AddressBook;
 import spleetwaise.address.model.AddressBookModel;
@@ -15,6 +14,7 @@ import spleetwaise.address.model.person.NameContainsKeywordsPredicate;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.testutil.Assert;
 import spleetwaise.address.testutil.EditPersonDescriptorBuilder;
+import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.commands.Command;
 import spleetwaise.commons.logic.commands.CommandResult;
 import spleetwaise.commons.logic.commands.exceptions.CommandException;
