+++ src/test/java/spleetwaise/commons/core/index/IndexTest.java
@@ -1,4 +1,4 @@
-package spleetwaise.address.commons.core.index;
+package spleetwaise.commons.core.index;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
