+++ src/test/java/spleetwaise/transaction/testutil/TypicalIndexes.java
@@ -1,6 +1,6 @@
 package spleetwaise.transaction.testutil;
 
-import spleetwaise.address.commons.core.index.Index;
+import spleetwaise.commons.core.index.Index;
 
 /**
  * A utility class containing a list of {@code Index} objects to be used in tests.
